feat: use class-oriented structure
